MS: changed hedgehog starting commands to the new standard

+++ bootstrap/src/main/java/org/unigrid/bootstrap/UpdateView.java
@@ -284,7 +284,7 @@ public class UpdateView implements UpdateHandler, Injectable, Initializable {
 		String onlineFileName = "";
 		System.out.println("is this working!!!!");
 		for (FileMetadata file : files) {
-			if (!file.isModulepath()) {
+			if (!file.getPath().getFileName().toString().contains("unigrid")) {
 				String s = file.getUri().toString();
 				String[] arr = s.split("/");
 				onlineFileName = arr[arr.length - 1];
@@ -304,7 +304,8 @@ public class UpdateView implements UpdateHandler, Injectable, Initializable {
 		String untarName = "";
 
 		for (FileMetadata file : files) {
-			if (!file.isModulepath()) {
+			if (file.getPath().toString().contains(".tar.gz")) {
+				System.out.println(file.getPath().toString());
 				String s = file.getUri().toString();
 				String[] arr = s.split("/");
 				untarName = arr[arr.length - 1];
@@ -323,7 +324,13 @@ public class UpdateView implements UpdateHandler, Injectable, Initializable {
 				if (a.isDirectory()) {
 					File[] unigrid = a.listFiles();
 					for (File b : unigrid) {
+						if (!b.exists()) {
+							break;
+						}
 						File[] bintar = b.listFiles();
+						if (bintar == null || bintar.length < 1) {
+							break;
+						}
 						for (File file : bintar) {
 							file.delete();
 						}
@@ -389,7 +396,7 @@ public class UpdateView implements UpdateHandler, Injectable, Initializable {
 		String untarName = "";
 
 		for (FileMetadata file : files) {
-			if (!file.isModulepath()) {
+			if (file.getPath().toString().contains(".zip")) {
 				String s = file.getUri().toString();
 				String[] arr = s.split("/");
 				untarName = arr[arr.length - 1];
